Feature: Movies CRUD endpoints, validation and DB connection (#1)
* build: installed dependencies, updated readme & tsconfig

Installed cors, dotenv, express-joi-validation, joi and mongoose. Also added better info to
readme.md. Finally updated tsconfig to match new requirements

BREAKING CHANGE: Installed new dependencies

* feat(mongodb/config.ts): create mongodb connection function

Created a connect function that takes no arguments and connects to the mongodb database using the
MONGODB_CONNECTION_STRING environment variable

* feat(mongodb/models): create mongodb models and index file to export them

Created a mongodb model for the "Movie" entity as well as the typescript type to map it when it gets
fetched from the db

* feat(mongodb/resolvers): create index file to export connect function

This commit includes an index file in the resolvers directory to export the connect function without
accessing the file directly

* feat(mongodb/resolvers/movies.ts): create movie resolvers to fetch/post movies to the database

This commit adds every function required to let the movies entity to interact with the mongodb
database through mongoose schemas.

* feat(controllers): create controllers for both the main route and movies routes

This commit creates every controller function required to handle the main route and CRUD routes for
the movies entity

* feat(validation/schemas): create schemas for endpoint validations using Joi

This commit adds the MovieValidationSchema, a Joi schema that will validate the body of both the
update and create requests (PUT and POST respectively). It also creates a separate validation schema
for the id passed in to the URL as a query parameter for all CRUD operations except for "GET all
movies

* feat(middlewares/validation): create validation middlewares using Joi schemas

This commit adds the middleware functions to handle both body and query-param validation for movies'
endpoints

* feat(routers): create routers to handle every movies endpoints

This commit integrates middlewares and controllers to handle movies and '/' endpoints.

* feat(app.ts, server.ts): update both app and server files

This commits add the final version of the app and server files to allow the API to run in the
desired port, connect to the database and log errors if any.

* feat(readme.md, controllers/main.ts): add prod link to readme and update main router msg

This commit added a production link to the project's README.md file as well as updated the message
for the API's main router.